poisson returns unsigned long

+++ src/rng.hpp
@@ -99,10 +99,10 @@ class RandomNumberGenerator {
             return n;
         }
 
-        // returns a real value sampled from a Poisson distribution with mean mu
-        inline double poisson(double mu) {
+        // returns an integer value sampled from a Poisson distribution with mean mu
+        inline unsigned long poisson(double mu) {
             if (mu <= 0) {
-                return 0.0;
+                return 0;
             }
             return gsl_ran_poisson(this->rgen_, mu);
         }
